requirements.txt
Cartopy==0.17.0
cdsapi==0.2.5
certifi==2019.11.28
cffi==1.13.2
cfgrib==0.9.7.4
cftime==1.0.4.2
chardet==3.0.4
Click==7.0
cycler==0.10.0
MarkupSafe==1.1.1
matplotlib==3.1.2
netCDF4==1.5.3
numpy==1.17.4
pandas==0.25.3
Pillow==6.2.1
pupygrib==0.5.0
pycparser==2.19
pyparsing==2.4.5
pyproj==1.9.3
pyshp==2.1.0
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
scipy==1.3.3
seaborn==0.9.0
Shapely==1.6.4.post2
six==1.13.0
urllib3==1.25.7
Werkzeug==0.16.0
xarray==0.14.1
